nemo json parser work started - basic system using rapidjson. Requires c++11 .. not sure about c++ 14. todo: Will hopefully save the data in a portable way. I have not ruled out creating a C wrapper for the function to be used in NeMo directly.

+++ src/IO/IOStack.h
@@ -183,7 +183,7 @@ void lPushParam(char *paramName);
  * @return
  */
 
-long lGetParam(int isArray, long *arrayParam);
+long long int lGetParam(int isArray, long *arrayParam);
 
 /**
  * Returns a specific global model configuration parameter.
